# perf: setup nunjucks environment on load rather than per request